move some tests to core
