+++ modules/api/src/test/scala/vinyldns/api/domain/record/RecordSetChangeSpec.scala
@@ -18,10 +18,9 @@ package vinyldns.api.domain.record
 
 import org.scalatest.{Matchers, WordSpec}
 import vinyldns.api.VinylDNSTestData
-import vinyldns.api.domain.dns.DnsConversions
 import vinyldns.core.domain.record.AData
 
-class RecordSetChangeSpec extends WordSpec with Matchers with VinylDNSTestData with DnsConversions {
+class RecordSetChangeSpec extends WordSpec with Matchers with VinylDNSTestData {
 
   "RecordSetChange" should {
     "toString" should {
